"adapt int or bool condition datatype"

+++ include/cuda/cuda_where.h
@@ -4,7 +4,7 @@
 
 namespace infini {
 void where_kernel(const float *inputX, const float *inputY,
-                  const int *condition, float *output, int nDims,
+                  const uint8_t *condition, float *output, int nDims,
                   infini::SmallArray inputXShape,
                   infini::SmallArray inputYShape,
                   infini::SmallArray conditionShape,
